+++ vizeo_theme/lib/private/enum.dart
@@ -1,4 +1,3 @@
-part of '../vizeo_theme.dart';
 
 enum VzTextType {
   typePrimary,