+++ vizeo_theme/lib/private/vz_strings.dart
@@ -1,4 +1,2 @@
-part of '../vizeo_theme.dart';
-
 
 const myFontFamily = 'DIN2014';