+++ vizeo_theme/lib/vizeo_theme.dart
@@ -1,20 +1,13 @@
 library vizeo_theme;
 
-import 'package:flutter/material.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/painting.dart';
-
-part 'versionning_project.dart';
-part 'widgets/vz_text.dart';
-part 'widgets/vz_chip.dart';
-part 'widgets/vz_button.dart';
-part 'widgets/vz_selectable_text.dart';
-part 'widgets/vz_form.dart';
-part 'widgets/vz_slider.dart';
-part 'widgets/vz_switch.dart';
-part 'themes/vz_light_theme.dart';
-part 'themes/vz_dark_theme.dart';
-part 'themes/vz_strings.dart';
-part 'themes/vz_colors.dart';
-part 'themes/enum.dart';
-part 'themes/color_utils_lib.dart';
+export 'widgets/vz_text.dart';
+export 'widgets/vz_chip.dart';
+export 'widgets/vz_checkbox.dart';
+export 'widgets/vz_button.dart';
+export 'widgets/vz_selectable_text.dart';
+export 'widgets/vz_form.dart';
+export 'widgets/vz_slider.dart';
+export 'widgets/vz_switch.dart';
+export 'themes/vz_light_theme.dart';
+export 'themes/vz_dark_theme.dart';
+export 'themes/vz_colors.dart';