+++ vizeo_theme/CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
 * Add to private repo
+## 0.0.18
+
+* themeData
+* principal component
+* convert part of -> to export for create some private part fo the lib